requirements.txt
# PyTorch with CUDA support for Linux
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
--index-url https://download.pytorch.org/whl/cu118

# Data visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0

# Common dependencies that might be useful
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.0.0

transformers
huggingface_hub[hf_xet]
